--- rafiki/src/App.css
@@ -1,90 +1,0 @@
-.learnWithUs{
-    display: flex;
-    justify-content: center;
-    gap: 200px;
-    margin-right: 100px;
-    
-
-
-
-}
-.heading{
-    text-align: center;
-    font-weight: 100px;
-    font-size: 50px;
-    color: rgb(17, 38, 65);
-    margin-right: 40px;
-}
-.text{
-    
-    font-size: 30px;
-    padding-top: 200px;
-}
-.image{
-    height:650px;
-    width:800px;
-}
-@media screen and (min-width:240px) and (max-width:800px)  {
-    .learnWithUs{
-        display: flex;
-         flex-wrap: wrap; 
-        gap: 5%;
-         /* margin-right: 100px; */
-         padding-left: 40px;
-    
-    }
-    .heading{
-        text-align: center;
-        font-family: Georgia, 'Times New Roman', Times, serif;
-        font-weight: 100px;
-        font-size: 30px;
-        color: rgb(21, 74, 143);
-        text-align: center;
-    }
-    .text{
-        font-family: Georgia, 'Times New Roman', Times, serif;
-        font-size: 16px;
-        padding-top: 50px; 
-        flex-wrap: wrap;
-        
-    }
-    .image{
-        height:300px;
-        width:350px;
-        /* margin-left: 30px;  */
-    }
-    
-
-}
-
-@media screen and (min-width:800px) and (max-width:1300px)  {
-    .learnWithUs{
-        display: flex;
-        /* flex-wrap: wrap; */
-        gap: 5%;
-         /* margin-right: 100px; */
-         padding-left: 40px;
-    
-    }
-    .heading{
-        text-align: center;
-        font-weight: 100px;
-        font-size: 30px;
-        color: rgb(12, 28, 49);
-        text-align: center;
-    }
-    .text{
-        
-        font-size: 25px;
-        padding-top: 50px; 
-        flex-wrap: wrap;
-        
-    }
-    .image{
-        height:400px;
-        width:450px;
-        /* margin-left: 30px;  */
-    }
-    
-
-}
