+++ rafiki/src/App.js
@@ -1,22 +1,11 @@
-import './App.css';
+import LearnWithUs from "./LearnWithUs";
 
-const LearnWithUs = ()=>{
+const App = ()=>{
   return(
-      <div>
-          <h1 className="heading">
-              Learn With Us 
-          </h1 >
-          <section className="learnWithUs">
-           <img src="/Images/students.png"alt="girl studying" className="image"></img>
-           <section className="text">
-             <p>Learning students are happy students.</p>
-             <p>Happy students make happy parents.</p>
-             <p>We give parents a hand while</p>
-             <p>giving students the tools to succeed. </p>
-           </section>
-           
-          </section>
-      </div>
+    <div>
+     <LearnWithUs/>
+      
+    </div>
   )
 }
-export default LearnWithUs;
+export default App;
