+++ rafiki/src/LearnWithUs/index.css
@@ -1,0 +1,204 @@
+.learnWithUs{
+    display: flex;
+    justify-content: center;
+    gap: 200px;
+    margin-right: 100px;
+    
+
+
+
+}
+.heading{
+    text-align: center;
+    font-weight: 30px;
+    font-size: 50px;
+    color: rgb(24, 69, 129);
+    margin-right: 40px;
+}
+.text{
+    
+    font-size: 28px;
+    padding-top: 200px;
+    padding-right: 50px;
+    padding-left: 80px;
+}
+.image{
+    height:650px;
+    width:850px;
+    padding-left: 80px;
+}
+@media screen and (min-width:300px) and (max-width:750px)  {
+    .learnWithUs{
+        display: flex;
+         flex-wrap: wrap; 
+        gap: 4%;
+         /* margin-right: 100px; */
+         padding-left: 80px;
+    
+    }
+    .heading{
+        text-align: center;
+        font-family: Georgia, 'Times New Roman', Times, serif;
+        font-weight: 100px;
+        font-size: 30px;
+        color: rgb(21, 74, 143);
+        text-align: center;
+    }
+    .text{
+        font-family: Georgia, 'Times New Roman', Times, serif;
+        font-size: 16px;
+        padding-top: 50px; 
+        flex-wrap: wrap;
+        padding-left: 5px;
+        padding-right: 10px;
+        text-align: center;
+        
+    }
+    .image{
+        height:300px;
+        width:350px;
+        margin-right: 70px; 
+    }
+    
+
+}
+
+@media screen and (min-width:200px) and (max-width:300px)  {
+    .learnWithUs{
+        display: flex;
+         flex-wrap: wrap; 
+        gap: 4%;
+         /* margin-right: 100px; */
+         padding-left: 85px;
+    
+    }
+    .heading{
+        text-align: center;
+        font-family: Georgia, 'Times New Roman', Times, serif;
+        font-weight: 100px;
+        font-size: 30px;
+        color: rgb(21, 74, 143);
+        text-align: center;
+    }
+    .text{
+        font-family: Georgia, 'Times New Roman', Times, serif;
+        font-size: 13px;
+        padding-top: 50px; 
+        flex-wrap: wrap;
+        padding-right: 2px;
+        padding-left: 2px;
+        
+        
+    }
+    .image{
+        height:250px;
+        width:300px;
+        margin-right: 70px;
+        
+    }
+    
+
+}
+
+@media screen and (min-width:750px) and (max-width:900px)  {
+    .learnWithUs{
+        display: flex;
+        /* flex-wrap: wrap; */
+        gap: 4%;
+        flex-direction: row;
+         /* margin-right: 100px; */
+         padding-left: 20px;
+    
+    }
+    .heading{
+        text-align: center;
+        font-weight: 100px;
+        font-size: 30px;
+        color: rgb(12, 28, 49);
+        text-align: center;
+    }
+    .text{
+        
+        font-size: 18px;
+        padding-top: 50px; 
+        flex-wrap: wrap;
+        padding-right: 5px;
+        padding-left: 10px;
+        
+        
+    }
+    .image{
+        height:300px;
+        width:400px;
+        margin-left: 30px;  
+    }
+    
+
+}
+
+@media screen and (min-width:900px) and (max-width:1050px)  {
+    .learnWithUs{
+        display: flex;
+        /* flex-wrap: wrap; */
+        gap: 4%;
+         /* margin-right: 100px; */
+         padding-left: 40px;
+    
+    }
+    .heading{
+        text-align: center;
+        font-weight: 100px;
+        font-size: 30px;
+        color: rgb(12, 28, 49);
+        text-align: center;
+    }
+    .text{
+        
+        font-size: 21px;
+        padding-top: 110px; 
+        flex-wrap: wrap;
+        padding-left: 10px;
+        
+    }
+    .image{
+        height:400px;
+        width:450px;
+        margin-left: 30px;
+    }
+}
+    
+
+    @media screen and (min-width:1050px) and (max-width:1300px)  {
+        .learnWithUs{
+            display: flex;
+            /* flex-wrap: wrap; */
+            gap: 3%;
+             /* margin-right: 100px; */
+             padding-left: 40px;
+        
+        }
+        .heading{
+            text-align: center;
+            font-weight: 100px;
+            font-size: 30px;
+            color: rgb(12, 28, 49);
+            text-align: center;
+        }
+        .text{
+            
+            font-size: 23px;
+            padding-top: 110px; 
+            flex-wrap: wrap;
+            padding-left: 130px;
+            padding-right: 70px;
+            
+            
+        }
+        .image{
+            height:450px;
+            width:550px;
+            margin-left: 80px;
+        }
+        
+
+}
