+++ rafiki/src/LearnWithUs/index.js
@@ -1,0 +1,21 @@
+import './index.css'
+const LearnWithUs = ()=>{
+    return(
+        <div>
+            <h1 className="heading">
+                Learn With Us 
+            </h1 >
+            <section className="learnWithUs">
+             <img src='Images/happystudents.jpg' alt="studying" className="image"></img>
+             <section className="text">
+               <p>Learning students are happy students.
+               Happy students make happy parents.
+               We give parents a hand while
+               giving students the tools to succeed. </p>
+             </section>
+             
+            </section>
+        </div>
+    )
+  }
+  export default LearnWithUs;
